chore: build services as a separate step
