+++ src/listr/tasks/startNodeTaskFactory.js
@@ -1,6 +1,4 @@
 const fs = require('fs');
-const { exec } = require('child_process');
-const { promisify } = require('util');
 
 const { Listr } = require('listr2');
 const { Observable } = require('rxjs');
@@ -13,7 +11,7 @@ const { NETWORK_LOCAL } = require('../../constants');
  * @param {waitForCorePeersConnected} waitForCorePeersConnected
  * @param {waitForMasternodesSync} waitForMasternodesSync
  * @param {createRpcClient} createRpcClient
- * @param {Docker} docker
+ * @param {buildServicesTask} buildServicesTask
  * @return {startNodeTask}
  */
 function startNodeTaskFactory(
@@ -21,11 +19,8 @@ function startNodeTaskFactory(
   waitForCorePeersConnected,
   waitForMasternodesSync,
   createRpcClient,
-  docker,
+  buildServicesTask,
 ) {
-  const execAsync = promisify(exec);
-  const followDockerProgress = promisify(docker.modem.followProgress.bind(docker.modem));
-
   /**
    * @typedef {startNodeTask}
    * @param {Config} config
@@ -67,94 +62,12 @@ function startNodeTaskFactory(
       },
       {
         title: 'Build services',
-        skip: (ctx) => ctx.skipFurtherServiceBuilds === true,
-        enabled: () => config.has('platform')
+        enabled: (ctx) => !ctx.skipBuildServices && config.has('platform')
           && (
             config.get('platform.dapi.api.docker.build.path') !== null
             || config.get('platform.drive.abci.docker.build.path') !== null
           ),
-        task: (ctx) => {
-          ctx.skipFurtherServiceBuilds = true;
-
-          const serviceBuildConfigs = [
-            {
-              name: 'Drive',
-              buildOptions: config.get('platform.drive.abci.docker.build'),
-              serviceName: 'drive_abci',
-            },
-            {
-              name: 'DAPI',
-              buildOptions: config.get('platform.dapi.api.docker.build'),
-              serviceName: 'dapi_api',
-            },
-          ];
-
-          const buildTasks = serviceBuildConfigs
-            .filter(({ buildOptions }) => buildOptions.path !== null)
-            .map(({
-              name,
-              buildOptions,
-              serviceName,
-            }) => ({
-              title: `Build ${name}`,
-              task: () => (
-                new Listr([
-                  {
-                    title: 'Build Docker image',
-                    task: async () => {
-                      const envs = config.toEnvs();
-
-                      await dockerCompose.build(envs, serviceName);
-                    },
-                  },
-                  {
-                    title: 'Update NPM cache',
-                    task: async () => {
-                      // Build node_modules stage only to access to npm cache
-                      const buildStream = await docker.buildImage({
-                        context: buildOptions.path,
-                        src: ['Dockerfile', 'docker/cache', 'package.json', 'package-lock.json'],
-                      }, {
-                        target: 'node_modules',
-                      });
-
-                      const output = await followDockerProgress(buildStream);
-
-                      const buildError = output.find(({ error }) => error);
-
-                      if (buildError) {
-                        throw new Error(buildError.error);
-                      }
-
-                      const {
-                        aux: {
-                          ID: nodeModulesImageId,
-                        },
-                      } = output.find(({ aux }) => aux && aux.ID);
-
-                      // Copy npm cache from node_modules stage image back to cache dir
-                      const container = await docker.createContainer({
-                        Image: nodeModulesImageId,
-                      });
-
-                      await Promise.all([
-                        execAsync(`docker cp ${container.id}:/root/.cache ${buildOptions.path}/docker/cache`),
-                        execAsync(`docker cp ${container.id}:/root/.npm ${buildOptions.path}/docker/cache`),
-                      ]);
-
-                      // Remove node_modules stage container and image
-                      await container.remove();
-
-                      const nodeModulesImage = docker.getImage(nodeModulesImageId);
-                      await nodeModulesImage.remove();
-                    },
-                  },
-                ])
-              ),
-            }));
-
-          return new Listr(buildTasks);
-        },
+        task: () => buildServicesTask(config),
       },
       {
         title: 'Start services',
